+++ mwparserfromhell/parser/__init__.py
@@ -26,6 +26,19 @@ modules: the :py:mod:`~.tokenizer` and the :py:mod:`~.builder`. This module
 joins them together under one interface.
 """
 
+class ParserError(Exception):
+    """Exception raised when an internal error occurs while parsing.
+
+    This does not mean that the wikicode was invalid, because invalid markup
+    should still be parsed correctly. This means that the parser caught itself
+    with an impossible internal state and is bailing out before other problems
+    can happen. Its appearance indicates a bug.
+    """
+    def __init__(self, extra):
+        msg = "This is a bug and should be reported. Info: {0}.".format(extra)
+        super(ParserError, self).__init__(msg)
+
+
 from .builder import Builder
 from .tokenizer import Tokenizer
 try:
@@ -35,15 +48,22 @@ except ImportError:
     CTokenizer = None
     use_c = False
 
-__all__ = ["use_c", "Parser"]
+__all__ = ["use_c", "Parser", "ParserError"]
 
 class Parser(object):
     """Represents a parser for wikicode.
 
     Actual parsing is a two-step process: first, the text is split up into a
-    series of tokens by the :py:class:`~.Tokenizer`, and then the tokens are
-    converted into trees of :py:class:`~.Wikicode` objects and
-    :py:class:`~.Node`\ s by the :py:class:`~.Builder`.
+    series of tokens by the :py:class:`.Tokenizer`, and then the tokens are
+    converted into trees of :py:class:`.Wikicode` objects and
+    :py:class:`.Node`\ s by the :py:class:`.Builder`.
+
+    Instances of this class or its dependents (:py:class:`.Tokenizer` and
+    :py:class:`.Builder`) should not be shared between threads.
+    :py:meth:`parse` can be called multiple times as long as it is not done
+    concurrently. In general, there is no need to do this because parsing
+    should be done through :py:func:`mwparserfromhell.parse`, which creates a
+    new :py:class:`.Parser` object as necessary.
     """
 
     def __init__(self):
@@ -65,6 +85,9 @@ class Parser(object):
 
         If *skip_style_tags* is ``True``, then ``''`` and ``'''`` will not be
         parsed, but instead will be treated as plain text.
+
+        If there is an internal error while parsing, :py:exc:`.ParserError`
+        will be raised.
         """
         tokens = self._tokenizer.tokenize(text, context, skip_style_tags)
         code = self._builder.build(tokens)